Refactor handlers into separate modules
